A3/requirements.txt
torch>=1.9.0
torchvision
torchtext
timm
transformers
huggingface_hub
safetensors
ftfy
regex
tqdm
filelock
opencv-python
scikit-image
scipy
pycocotools
h5py
mmcv-full==1.3.12
mmsegmentation==0.17.0
matplotlib
Pillow
pillow-simd
requests
pandas
git+https://github.com/salesforce/BLIP.git